requirements.txt
opencv_python==4.11.0.86
PyYAML==6.0.2
PyYAML==6.0.2
ultralytics==8.3.67
